Big numbers to words

+++ Bloodrush/Assets/Scripts/Backend/Upgrade.cs
@@ -4,19 +4,19 @@ using System.Collections;
 
 public class Upgrade : MonoBehaviour {
             
-    public void BuyBPM(float BPMAmount, int CalCost, BeatManager bm)
+    public void BuyBPM(float BPMAmount, ulong CalCost, BeatManager bm)
     {
         bm.TimeBetweenBeats(BPMAmount);
         ScoreManager.DecreaseCalories(CalCost);
     }
 
-    public void BuyBC(int BCAmount, int O2Cost)
+    public void BuyBC(ulong BCAmount, ulong O2Cost)
     {
         ScoreManager.IncreaseMult(BCAmount);
         ScoreManager.DecreaseOxygen(O2Cost);
     }
 
-    public void BuyCAL(int CalAmount, int O2Cost, int O2pBCost)
+    public void BuyCAL(ulong CalAmount, ulong O2Cost, ulong O2pBCost)
     {
         ScoreManager.IncreaseCALMult(CalAmount);
         ScoreManager.DecreaseOxygen(O2Cost);
